# optimize cast for same decimal type and same scale