# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

we are optimizing the TPCH query, and find some decimal cast is heavy such TPCH-1.



In the Sum Agg, the sum decimal type is (precision+10, scale). The scale is not changed, the array don't need to be converted.



# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

add two micro benchmark, and get the result:

```
cast decimal128 to decimal128 512 with same scale
                        time:   [213.91 ns 214.46 ns 215.11 ns]
                        change: [-96.673% -96.654% -96.628%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe

cast decimal256 to decimal256 512 with same scale
                        time:   [212.81 ns 213.51 ns 214.28 ns]
                        change: [-98.529% -98.524% -98.519%] (p = 0.00 < 0.05)
                        Performance has improved.
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->